requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
flet==0.25.1
flet-desktop==0.25.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
itsdangerous==2.2.0
oauthlib==3.2.2
pycparser==2.22
python-dotenv==1.0.1
repath==0.9.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
